# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Verity API",
    "description": "MVP monolítico modular para gestión documental con IA, aprobaciones humanas, y agente conversacional Veri.",
    "version": "1.0.0",
    "contact": {
      "name": "Verity Team"
    }
  },
  "servers": [
    {
      "url": "https://verity-api-{env}.a.run.app",
      "description": "Google Cloud Run",
      "variables": {
        "env": {
          "default": "dev",
          "enum": ["dev", "staging", "prod"]
        }
      }
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development"
    }
  ],
...
  "paths": {
...
    "/documents/{document_id}": {
      "get": {
        "tags": ["documents"],
        "summary": "Get document metadata",
        "operationId": "getDocument",
        "security": [{"bearerAuth": []}],
        "x-feature-flag": "FEATURE_DOCUMENTS",
        "parameters": [
          {"$ref": "#/components/parameters/DocumentId"}
        ],
        "responses": {
          "200": {
            "description": "Metadata del documento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "503": {"$ref": "#/components/responses/FeatureDisabled"}
        }
      },
      "delete": {
        "tags": ["documents"],
        "summary": "Delete document",
        "operationId": "deleteDocument",
        "security": [{"bearerAuth": []}],
        "x-feature-flag": "FEATURE_DOCUMENTS",
        "x-roles": ["admin"],
        "parameters": [
          {"$ref": "#/components/parameters/DocumentId"}
        ],
        "responses": {
          "204": {"description": "Documento eliminado"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "503": {"$ref": "#/components/responses/FeatureDisabled"}
        }
      }
    },
...
    "/logs": {
      "get": {
        "tags": ["logs"],
        "summary": "List logs (admin only, read-only)",
        "description": "Lista logs de la aplicación. SOLO LECTURA, sin filtros arbitrarios para evitar leaks. Requiere rol admin.",
        "operationId": "listLogs",
        "security": [{"bearerAuth": []}],
        "x-feature-flag": "FEATURE_LOGS",
        "x-roles": ["admin"],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
            },
            "description": "Filtro fijo por nivel (único filtro permitido)"
          },
          {
            "name": "since",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Logs desde esta fecha"
          },
          {"$ref": "#/components/parameters/PageSize"},
          {"$ref": "#/components/parameters/PageToken"}
        ],
        "responses": {
          "200": {
            "description": "Lista de logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogListResponse"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "503": {"$ref": "#/components/responses/FeatureDisabled"}
        }
      }
    },
    "/logs/stream": {
      "get": {
        "tags": ["logs"],
        "summary": "Stream logs via SSE (admin only, read-only)",
        "description": "Streaming de logs en tiempo real estilo terminal. SOLO LECTURA. Requiere rol admin.",
        "operationId": "streamLogs",
        "security": [{"bearerAuth": []}],
        "x-feature-flag": "FEATURE_LOGS",
        "x-roles": ["admin"],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]
            },
            "description": "Filtro fijo por nivel mínimo"
          }
        ],
        "responses": {
          "200": {
            "description": "Stream de logs (SSE)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-Sent Events con logs formateados"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "503": {"$ref": "#/components/responses/FeatureDisabled"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Supabase Auth. Include in header: Authorization: Bearer <token>"
      }
    },
    "parameters": {
      "DocumentId": {
        "name": "document_id",
        "in": "path",
        "required": true,
        "schema": {"type": "string", "format": "uuid"}
      },
...
      "PageSize": {
        "name": "page_size",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "PageToken": {
        "name": "page_token",
        "in": "query",
        "schema": {"type": "string"},
        "description": "Token for next page"
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing JWT",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
...
      "FeatureDisabled": {
        "description": "Feature Disabled - Module is turned off via feature flag",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "examples": ["UNAUTHORIZED", "NOT_FOUND", "FEATURE_DISABLED", "VALIDATION_ERROR"]
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "additionalProperties": true,
                "description": "Additional error context"
              },
              "request_id": {
                "type": "string",
                "format": "uuid",
                "description": "Request ID for debugging"
              }
            }
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total_count": {"type": "integer"},
          "page_size": {"type": "integer"},
          "next_page_token": {"type": "string", "nullable": true},
          "has_more": {"type": "boolean"}
        }
      },
      "DocumentResponse": {
        "type": "object",
        "required": ["id", "display_name", "gemini_uri", "mime_type", "size_bytes", "status", "created_at", "created_by"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "display_name": {"type": "string"},
          "gemini_uri": {"type": "string", "description": "Gemini File API URI"},
          "mime_type": {"type": "string"},
          "size_bytes": {"type": "integer"},
          "status": {
            "type": "string",
            "enum": ["processing", "ready", "failed"]
          },
          "metadata": {"type": "object", "additionalProperties": true},
          "created_at": {"type": "string", "format": "date-time"},
          "created_by": {"type": "string", "format": "uuid"}
        }
      },
...
      "LogEntry": {
        "type": "object",
        "required": ["timestamp", "level", "message"],
        "properties": {
          "timestamp": {"type": "string", "format": "date-time"},
          "level": {"type": "string", "enum": ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]},
          "message": {"type": "string"},
          "logger": {"type": "string"},
          "request_id": {"type": "string", "format": "uuid", "nullable": true},
          "user_id": {"type": "string", "format": "uuid", "nullable": true},
          "extra": {"type": "object", "additionalProperties": true}
        }
      },
      "LogListResponse": {
        "type": "object",
        "required": ["items", "meta"],
        "properties": {
          "items": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/LogEntry"}
          },
          "meta": {"$ref": "#/components/schemas/PaginationMeta"}
        }
      },
...
    }
  },
  "security": [
    {"bearerAuth": []}
  ]
}